Refactor project structure
- remove build/file system artifacts and add .gitignore
- use pyproject.toml and flit for project structure
- change package structure to allow for a simple `import libsoni`
- change demo notebooks accordingly
