+++ .gitignore
@@ -1,0 +1,29 @@
+# development stuff SJS
+notebooks
+scratch
+
+# MacOS artifacts
+.DS_Store
+
+# Python/Jupyter artifacts
+__pycache__
+.ipynb_checkpoints
+.Python
+.pytest_cache
+build/
+develop-eggs/
+dist/
+downloads/
+eggs/
+.eggs/
+lib/
+lib64/
+parts/
+sdist/
+var/
+wheels/
+share/python-wheels/
+*.egg-info/
+.installed.cfg
+*.egg
+MANIFEST
