--- environment_libsoni.yml
@@ -1,28 +1,0 @@
-name: libsoni
-
-channels:
-  - conda-forge
-  - defaults
-
-dependencies:
-  - python >=3.7.0, <4.0.0
-  - pip >=19.2.0, <20.0.0
-  - numpy >=1.17.0, <2.0.0
-  - scipy >=1.3.0, <2.0.0
-  - matplotlib >=3.1.0, <4.0.0
-  - ipython >=7.8.0, <8.0.0
-  - jupyter >=1.0.0, <2.0.0
-  - pandas >=1.0.0, <2.0.0
-  - scikit-learn >=0.21.0, <1.0.0
-  - ffmpeg >=2.8.0, <3.0.0
-  - numba >=0.51.0, <1.0.0
-  - libsndfile >=1.2.2
-  - pip:
-    - mir-eval >=0.5.0, <1.0.0
-    - nbstripout >=0.3.0, <1.0.0
-    - music21 >=5.7.0, <6.0.0
-    - pretty_midi >=0.2.0, <1.0.0
-    - librosa >=0.8.0, <1.0.0
-    - pysoundfile >=0.9.0, <1.0.0
-    - libfmp
-    - synctoolbox
