--- libsoni.egg-info/SOURCES.txt
@@ -1,22 +1,0 @@
-LICENSE
-README.md
-setup.py
-libsoni/__init__.py
-libsoni.egg-info/PKG-INFO
-libsoni.egg-info/SOURCES.txt
-libsoni.egg-info/dependency_links.txt
-libsoni.egg-info/requires.txt
-libsoni.egg-info/top_level.txt
-libsoni/core/__init__.py
-libsoni/core/chroma.py
-libsoni/core/f0.py
-libsoni/core/methods.py
-libsoni/core/pianoroll.py
-libsoni/core/spectrogram.py
-libsoni/core/tse.py
-libsoni/util/__init__.py
-libsoni/util/utils.py
-tests/test_f0.py
-tests/test_methods.py
-tests/test_pianoroll.py
-tests/test_spectrogram.py
