+++ libsoni/__init__.py
@@ -1,0 +1,16 @@
+"""libsoni
+
+'A Python toolbox for sonifying music annotations and feature representations
+"""
+from .version import version as __version__
+
+# import core functionalities into global namespace
+from .core.chroma import *
+from .core.f0 import *
+from .core.methods import *
+from .core.pianoroll import *
+from .core.spectrogram import *
+from .core.tse import *
+
+# keep utility functions in a separate sub-namespace
+from . import utils
