+++ libsoni/utils.py
