+++ libsoni/version.py
@@ -1,0 +1,1 @@
+version = "1.1.0"
